move stereo odom class outside its main.cpp so it can be reused
